requirements.txt
google-generativeai
haxor
html2text
requests